src/fund_management.rs
#![no_std]

use soroban_sdk::{contract, contractimpl, Address, Env, Symbol};

use crate::types::{Transaction, TransactionState};
use crate::errors::EscrowError;
use crate::audit::AuditModule;

/// Module for managing funds in escrow.
pub struct FundManagementModule;

#[contractimpl]
impl FundManagementModule {
    /// Deposits funds into escrow for a given transaction.
    pub fn deposit_funds(env: Env, transaction: &Transaction) -> Result<(), EscrowError> {
        // Ensure the transaction is in the correct state for deposit.
        if transaction.state != TransactionState::Setup {
            return Err(EscrowError::InvalidTransactionState);
        }

        // Transfer funds from the buyer to the escrow contract.
        let buyer = transaction.buyer;
        let amount = transaction.amount;
        buyer.require_auth();
        
        // Assuming a token client is available for transferring tokens.
        let token_client = token::Client::new(&env, &transaction.token);
        token_client.transfer(&buyer, &env.current_contract_address(), &amount);

        // Update the transaction state to indicate funds have been deposited.
        let mut updated_transaction = transaction.clone();
        updated_transaction.state = TransactionState::Deposit;
        env.storage().set(&transaction.id, &updated_transaction);

        // Record the deposit action for audit purposes.
        AuditModule::record_action(&env, &transaction.id, "deposit_funds");

        Ok(())
    }

    /// Releases funds from escrow to the seller upon fulfillment of conditions.
    pub fn release_funds(env: Env, transaction: &Transaction) -> Result<(), EscrowError> {
        // Ensure the transaction is in the correct state for release.
        if transaction.state != TransactionState::Deposit {
            return Err(EscrowError::InvalidTransactionState);
        }

        // Transfer funds from the escrow contract to the seller.
        let seller = transaction.seller;
        let amount = transaction.amount;
        
        // Assuming a token client is available for transferring tokens.
        let token_client = token::Client::new(&env, &transaction.token);
        token_client.transfer(&env.current_contract_address(), &seller, &amount);

        // Update the transaction state to indicate funds have been released.
        let mut updated_transaction = transaction.clone();
        updated_transaction.state = TransactionState::Complete;
        env.storage().set(&transaction.id, &updated_transaction);

        // Record the release action for audit purposes.
        AuditModule::record_action(&env, &transaction.id, "release_funds");

        Ok(())
    }
}
